Moved custom facility icon content to its own feature file.
